line 24 must repair
